Move some ARTDeviceIdentityTokenDetails methods into private header
I’m making this change right now because I want to replace the ARTLog in
these method signatures with ARTInternalLogHandler, the latter being a
type private to the SDK.

However, this is a change we should be making anyway — these methods
should not be part of the public interface of the SDK. I am happy to
consider this a non-breaking change (i.e. one that can be released in a
minor version update).

(We already have a broader issue — #1497 — for cleaning up the public
interface of the SDK.)