Encapsulate connection state change request metadata
In the upcoming implementation of #1431, I’m going to add further
(private) information to the ARTConnectionStateChange object emitted by
an ARTRealtime instance. This information needs to get from the place
that it is generated, all the way to the ARTRealtime method that emits
the state change. This method is currently named
transition:withErrorInfo:, and if we decide we want to emit further
information then this method will need to gain new parameters. Hence,
everything that (directly or indirectly) calls this method will need to
be updated to accept these new parameters (or to provide overrides that
only accept a subset of these parameters).

So, instead, bundle up all of the metadata related to a connection state
change request into a single object, so that as we expand this metadata
we don’t need to keep changing method signatures (except in the
initializers for this metadata, which can easily gain convenience
initializers to avoid having to change existing code if we wish).

(Further down the line, we may also wish to add some sort of request ID
to this metadata, which can be logged at the point where the request is
generated and the point where it is handled, so that our logs give us
richer information about why, for example, the library emitted a given
state change.)