Encapsulate channel state change request metadata
Same motivation as in f2bcb2d, but for channel. We’re currently misusing
ARTStatus for this purpose. But ARTStatus is a general purpose object
representing “the status of an operation”, and we’re abusing it by
giving it properties that only relate to a channel state change request
(see storeErrorInfo).

As part of this change, we remove the inappropriate storeErrorInfo
property from ARTStatus.

The convenience initializers of ARTChannelStateChangeMetadata are copied
from those of ARTStatus, to be able to make minimal changes to existing
code. I’m not convinced that the state:errorInfo: initializer should
always be setting storeErrorInfo to YES (feels like in the nil errorInfo
case it should behave the same as the state: initializer), but that’s
outside the scope of this work; created #1595 to look into it.