Encapsulate the request to attach a channel
This is similar to 5df1a9d, but one layer of indirection further out.
The aim (needed for the upcoming implementation of #1431) is for the
ARTRealtimeChannelSuspended case in ARTRealtimeChannelInternal’s
transition:withMetadata: method to be able to add metadata to the
emitted channel state change object. (To achieve this, we still need to
update -reattachWithReason: to accept a metadata object, which we’ll do
next.)

The convenience initializers of ARTAttachRequestMetadata match the
removed overloads of internalAttach.