Remove ARTCipherParams.logger property
Since this class is constructed by the user of the library and its
public initializer doesn’t accept a logger, this property is always nil,
so I’m not sure how it was intended to be used (added in ea77eef).

Our options are to get rid of the property or to add a new public
initializer to allow the user to populate it. Given that we’re only
logging one error message, which anyway subsequently appears inside a
thrown exception, it seems reasonable to just get rid of it.