Remove +[ARTDefault setRealtimeRequestTimeout]
We shouldn’t be using global state to control the behaviour of our
tests; it can lead to one test having an unintended impact on another.
Instead, let’s control it using a TestClientOptions property.

We should make this change for the other global state, too. I’ve
created #1601 for this.