Add a helper class for gathering data in tests
I’m going to refactor a test and wish to write it in a style where first
we gather data and then make assertions about it (as opposed to mixing
the gathering and the assertions), which allows us to fit the tests into
a Given / When / Then structure and hopefully make them more readable.

I think that this class will be helpful for doing so.