Introduce macros for logging
For #1643 (emit file name and line number with every logged message), it
would be good to have a way to do this without needing to pass __FILE__
and __LINE__ at every log statement call site. The only way to do this
is by using macros.

In an upcoming commit we’ll update the codebase to use these macros for
logging.

The implementation of ARTInternalLog’s `-logWithLevel:format:` is copied
from the existing "shorthand" methods.

By the way, the ##__VA_ARGS__ syntax is described in [1]. It’s used to
delete a trailing comma when the variable arguments list is empty. I saw
it being used in the `os` module’s log.h file to define the os_log
macro.

[1] https://gcc.gnu.org/onlinedocs/cpp/Variadic-Macros.html