Make ARTLog* macros emit file and line number information
There are currently two versions of the logging methods — one that
accepts file and line information and one that doesn’t. This will be
resolved when we remove the ability to log a message without file and
line information (for #1643). This also applies to the slightly dodgy
use of -1 in MockInternalLogBackend to communicate an absent integer
to Objective-C.