Split ARTInternalLog in two
ARTInternalLog already exposes functionality that is only accessible
from Objective-C (variadic methods). As part of #1643, we’re going to
add further such functionality (macros).

All of this functionality will ultimately end up calling a single "log
this string plus some metadata" method, and it would be good to be able
to test the Objective-C wrapper functionality separate from testing this
method.

Hence we split ARTInternalLog into two — a façade object, which provides
the Objective-C convenience functionality, and a core object, which
exposes the funnel "log this" method (which, as a bonus, is fully
mockable and testable using only Swift).