Remove code for accepting self-signed TLS certificates in tests
The commit history isn’t too helpful about why this code exists, but
looks like it was perhaps part of the ably-cocoa codebase itself, and
then moved to the test suite? Not sure. But removing it makes no
difference.

I asked on Slack [1] why this code might exist and it was mentioned that
running Realtime locally uses self-signed certs, but that’s not
something I believe we ever do as part of testing ably-cocoa. If it
turns out we ever do need it, we can restore it, properly documented.

[1] https://ably-real-time.slack.com/archives/C030C5YLY/p1681418354988069