Fix wait logic in PushAdminTests.setUp
The current logic assumes that the deviceRegistrations.save calls will
complete in the order that they are started (and hence, that when the
final device in the list has been registered, all of the devices have
been registered). This is not correct, and means that sometimes we end
up trying to create a channel subscription for a device that has not
been registered yet, causing the channelSubscriptions.save call to fail.