Use simpler implementation of continuous clock
Whilst working on aef5062, I noticed that the implementation of
+[ARTTime timeSinceBoot] seems to be sensitive to user changes in the
device clock. It also doesn’t seem to be thread-safe, making use of
global variables without any synchronisation.

Here we replace it with an implementation that uses an OS-provided
continuous clock, and introduce an API for dealing with the related
instants in time.

Using an object to represent fallbackRetryExpiration, instead of a
double, also gives us a clearer way of identifying the scenario where
this value has not been set, instead of using the arbitrary value of
zero.

The approach taken here also gives us a path to follow if in the future
we decide that we want to mock out some time-related functionality in
our tests (i.e. by turning ARTContinuousClock into a protocol and
injecting a mock instance).