Fix premature discarding of fallback host
Running RestClientTests on my development machine reveals multiple test
failures which are caused by the fact that the SDK incorrectly believes
that the timeout described by RSC15f has elapsed.

The underlying cause is that we're not initializing the value pointed to
by the `oldlenp` pointer passed to `sysctl`.

These tests are not failing in CI, which leads me to wonder if the
symptoms only manifest on ARM devices (e.g. my development machine). We
don't have an quick way of running our unit tests on actual iOS devices
so I haven't been able to confirm if they're affected.

Resolves #1683.