Throw instead of crash on sandbox app creation error
Sometimes the request to create an app in the sandbox times out, causing
the test suite to crash with a fatalError. This causes the test suite to
need to be restarted, and we have to go looking through crash reports to
find the cause of the crash. Instead, just treat this timeout as a test
failure.