Implementation for RTN20c with a workaround by avoiding immidiate restart of the connection that have been just started.
There is no certain way to detect whether reachability callback was called because you've just set it up, or because it really has detected reachability of the host. So we rely on a short timeout here - if the callback was called shortly after the setting up reachability, then reconnection shouldn't be done. Whereas if it was called after the certain delay (a threshold) - than it means you are in the middle of the dead-end connection attemt and it should be restarted.

See https://github.com/ably/ably-cocoa/issues/1713 for more details.