Fix test__011__Connection__should_disconnect_the_transport_when_no_activity_exist
I noticed that this is failing when I run it locally, and also
occasionally in CI [1].

The failure was introduced by cb7af9d. Before that commit, we were only
changing the global realtimeRequestTimeout value to 0.5 _after_
instantiating the ARTRealtime object, which meant that the client would
use the default timeout of 10s when initially waiting to receive a
CONNECT ProtocolMessage. However, cb7af9d unintentionally lowered this
waiting period to 0.5s, which, depending on factors such as network
conditions, may be too short. Since, as of cb7af9d, the
realtimeRequestTimeout used by a given ARTRealtime object is no longer
mutable, we need to pick a value that is sufficiently long to for us to
be able to expect to receive a CONNECT ProtocolMessage, but still fairly
short so that the test doesn’t run for a long time. 1.5s seems to meet
these needs.

[1] https://test-observability.herokuapp.com/repos/ably/ably-cocoa/test_cases/22dfe7d6-4289-4002-b135-449d179e7b9d