Fix instructions on embedding Ably framework
The current instructions re not embedding the frameworks in a macOS app
cause the app to exit upon launch with a linker error because one of our
frameworks can’t be found.

These instructions were added in 9691ebc, as part of the PR for adding
Swift Package Manager support, and there’s no explanation of the
motivation for this change or how it was related to that work.

My best guess is that it was probably something to do with the fact that
we previously used to embed our dependency frameworks inside
Ably.xcframework. However, this caused duplicate framework issues if we
also tried to embed the dependency frameworks inside the application
too. This is the issue described by #1594, which was resolved in 9a1e84c
by no longer embedding the dependency frameworks inside Ably.framework.

So, I’m removing these instructions, hence reverting 9a1e84c, and have
replaced them with language taken from the Carthage quickstart guide
[1].

Resolves #1183.

[1] https://github.com/Carthage/Carthage/blob/187a78c62811d3d75a9b1d41bfaeff708936125d/README.md#L57