Update connection test to handle bug #1782
This bug describes how the library will sometimes perform an extra retry
when it should transition to the SUSPENDED state. This can cause
test__059 to fail depending on the exact chosen value of
connectionStateTtl. In #1431 I will be changing the value of
connectionStateTtl that this test uses, and the value I want to use does
indeed cause the test to fail. So, adjust the test to handle the case
where an extra retry occurs. We should remove this workaround once we
fix #1782.