Removed `ARTPushRegistererDelegate.shouldRequestOtherDeviceTokensForAblyPush` delegate method.
Turned out it's not needed and `didActivateAblyPush:` works for the purpose of requesting another token just fine. I can't now recall in memory what exactly triggered me to think otherwise, but after a bunch of testing (in tests and manually) I couldn't reproduce any issues with this callback. Despite the fact that in the code of different `ARTPushActivationState`(s) the `callActivatedCallback:` is being called before setting a new state for the state machine, events produced by it (`GotPushDeviceDetails` in this case) are scheduled on the state machine's queue in a proper order (especially if there is an async call to get a new device token). And even if it wouldn't, the state machine is indeferent to it, since it has a pending events queue, and "inappropriate" event will be handled later.