Add a link to the GitHub job in observability server upload
Currently, the observability server UI links to the workflow run attempt
that the upload corresponds to. This is fine in the case where the
workflow only contains a single job, as it currently does in ably-cocoa.
But in the case where there’s more than one job in a workflow (e.g. in
the case of a matrix build, which we intend to switch ably-cocoa to
using in #1311), there’s no way to tell which job an upload corresponds
to. So, include this information in the upload.

Copied from test-observability-action commit d4d651c.