Add method for creating wrapper SDK proxy client
Part of #2020; that is, implementing ADR-117 [1]. This adds a
-createWrapperSDKProxyWithOptions: method to ARTRealtime. This method is
only to be called by Ably-authored wrapper SDKs and accepts a set of
agents describing the wrapper SDK. It returns a proxy object that wraps
the underlying ARTRealtime instance.

The intention is that when certain methods are called on this proxy, the
proxy will insert information that allows us to track the usage of the
wrapper SDK. I'll implement this behaviour in subsequent commits.

TODO some smoke tests of the API, and check that it's sharing state with
underlying realtime

[1] https://ably.atlassian.net/wiki/spaces/ENG/pages/3413016589/ADR-117+Tracking+wrapper+SDK+usage+continued