Remove initializers out of client protocols
Preparation for introducing a new class that conforms to these protocols
but which won't have these initializers. Let's use these protocols just
to list the functionality provided by an already-initialized client.

(To do this in a backwards-compatible way I've had to not actually
remove anything from the existing protocols, but rather introduce new
ones, which is a bit messy. Have created #2019 for us to sort this out
in our next major release.)