Create wrapper SDK proxy channel
Part of #2020; that is, implementing ADR-117 [1]. There are no
functional changes here; it's preparation for adding the `agent` channel
param when a channel is fetched via an ARTWrapperSDKProxyRealtime.

I initially implemented this with internal storage of proxy instances to
ensure that there be only one proxy instance per underlying channel.
However, it wasn't really clear to me _why_ I did this, other than for
some sort of symmetry with the underlying ARTRealtimeChannels behaviour.
So it didn't really seem worth the time of trying to figure out the
right behaviour. So I decided to not implement this behaviour in the
end, instead creating a new proxy instance each time -get:… is called.

[1] https://ably.atlassian.net/wiki/spaces/ENG/pages/3413016589/ADR-117+Tracking+wrapper+SDK+usage+continued