Revert 26d9bf7
That change, which moved the `presence` property from the
ARTRealtimeChannel class to ARTRealtimeChannelProtocol, changed the type
of the property from ARTRealtimePresence to
id<ARTRealtimePresenceProtocol>. This is a breaking change for anybody
who was relying on the property returning the concrete type, one such
example being our Asset Tracking SDK. So, we undo it.

The change made in 26d9bf7 was triggered by us noticing, whilst working
on the Chat SDK — which does its best to avoid referring to ably-cocoa's
concrete types — that when you have an ably-cocoa concrete type
ConcreteA which implements a protocol ProtocolA, and where ConcreteA has
a public method or property `someMethod` which returns an instance of
another concrete type ConcreteB which happens to implement a protocol
ConcreteB, then `someMethod` is always defined directly on ConcreteA,
not on ProtocolA. We wondered "why can't these methods be on ProtocolA
instead?", hence 26d9bf7. But the first reason is that we can't because
in order to do this it makes sense for you to also change the the return
type of `someMethod` from ConcreteB to ProtocolB, which is a breaking
change as we've seen here. But also I'm not sure whether it'd work even
if we accepted the breaking change; for example, I don't know if you can
mark a protocol as Sendable in Objective-C (haven't looked into it since
we currently have no intention of making the breaking change, but one to
investigate if we consider it in the future).

Resolves #2003.