Additional tests for RTN17b
I've decided to put this in a dedicated commit, otherwise it makes harder to review major changes in tests. Also removed redundant `options.fallbackHostsUseDefault` check since it false by default and to avoid deprecation message on top of the tests where this property is used, except the one where it's actually being tested.