Remove the SPM module map
This module map is not actually being used; per [1] it should be named
`module.modulemap`.

I tried renaming it and there are a bunch of issues to fix, which would
be a non-trivial task. So let's just get rid of this file for now, until
we actually need it (e.g. if we do #2034 and switch to SPM for running
tests, in which case the tests would need to be able to access the
Private submodule).

[1] https://github.com/swiftlang/swift-package-manager/blob/main/Documentation/Usage.md#creating-c-language-targets