Add a first attempt at the public API
Based on that of ably-js at f6fbe35.

This only is a first attempt so that we have _something_ to implement,
and I didn't want to get bogged down by making lots of small decisions
right now. Have created #4 for revisiting for improved Swiftiness.

The documentation comments are copied from JS. Used Cursor to convert it
to Swift syntax and then made some minor edits myself.