Add a basic plugins mechanism
This provides an initial implementation of the plugins mechanism
described in ADR-128 [1].

It adds a new SPM library product called AblyPlugin, which Ably-authored
plugins can import to:

- access private ably-cocoa functionality (the only such functionality
  currently exposed being access to a channel's logger; further
  functionality will be added as we develop the LiveObjects plugin)
- access protocols that ably-cocoa requires plugins to implement

It also adds the `plugins` client option, which is the mechanism that
users will use to pass plugins when initializing the SDK.

As a concrete example of a plugin, it adds the ability to provide a
LiveObjects plugin. See the corresponding LiveObjects plugin PR [2] to
see how to make use of this mechanism. We'll expand the library's
interactions with the LiveObjects plugin as we develop the LiveObjects
plugin.

[1] https://ably.atlassian.net/wiki/spaces/ENG/pages/3838574593/ADR-128+Plugins+for+ably-cocoa+SDK
[2] https://github.com/ably/ably-cocoa-liveobjects-plugin/pull/6