Add a library for testing plugins
Add a new SPM library product called `AblyPluginTesting`, which is to be
used by the test suite of Ably-authored plugins. It will expose some of
ably-cocoa's private APIs as needed by tests.

I'm fairly sure that we'll need something like this, having looked at
the usage of private APIs in the ably-js LiveObjects tests. If it turns
out we don't, we can remove it later.

There is no actual API here yet; I'll add it later as we start
developing the LiveObjects plugin.

Not yet decided whether this API needs to be stable; if it we make it
that we only ever run the LiveObjects tests against a fixed version of
ably-cocoa then it doesn't need to be. Can think about this later.

Note that I'm not attempting to reuse any of the test helpers that exist
in the ably-cocoa test suite, because they're written in Swift and also
make heavy use of Nimble.