Switch JSONValue's `number` case to use NSNumber
We need to be able to communicate to ably-cocoa's MessagePack encoder
that it should encode certain values (e.g. enums) as integers, not
doubles. This seems the quickest way of doing that. Don't _love_ it,
because NSNumber is not a very fun type to work with or make assertions
about. Will probably revisit when we introduce further
MessagePack-specific types (e.g. binary data).