Encode and decode data of WireObject
Based on [1] at 2e975cb.

This was implemented by updating the code to reflect the internal API
that I wanted to exist, and then asking Cursor to implement the rules of
the spec and to write tests. I then edited the generated code to
simplify it a bit and add things like @spec annotations and other
explanatory comments.

I wanted some round-trip tests that go through the Realtime backend but
decided to leave them until later once I have a bit more knowledge of
the LiveObjects protocol; have created #17.

[1] https://github.com/ably/specification/pull/335