Implement the OBJECT_SYNC specification
Based on [1] at dd25dca.

The development approach here was roughly the following: I decided the
internal types and interfaces that I wanted to exist, and some of the
implementation details (e.g. the mutations that DefaultRealtimeObjects
performs during a sync sequence). I then asked Cursor to fill in the
most of the implementation and tests (providing it with the relevant
spec text). I have done some tweaking of the code that Cursor generated,
but that doesn't mean that all of the code that's here is exactly the
same as code that I would write, and I'm going to leave it that way.

I have looked through all (and changed some) of the tests that Cursor
generated, and in doing so also fixed the spec conformance tags (because
it got those very wrong), so I am _fairly_ confident that the tests are
testing what they claim to test. But I have not looked at them in minute
detail (in particular, the exact data created by the canned data
factories that it created). I also think that tests are one of those
things where there's not always one right way to write them, so I'm not
going to get hung up on "could it have done this test differently?" or
"are the tests consistent with each other?" at the moment.

There are some outstanding questions on the spec PR at the moment; have
implemented to best of my understanding, and will update later once
those are answered.

Not yet implemented:

- Checking of channel modes before performing operations (RTO2 and the
  points that refer to it); there's an outstanding question about this
  at [2]

A note re the MutableState pattern that I've used here — done this so
that the class can essentially have mutating instance methods which can
call each other without having to worry about whose responsibility it is
to acquire the mutex.

[1] https://github.com/ably/specification/pull/333
[2] https://github.com/ably/specification/pull/333/files#r2152297442