Add concurrency annotations to the public API
Mark most of its types as Sendable, and callbacks as `sending`.

I was going to defer this until later but when I started writing
integration tests I realised that it's quite hard to actually use the
SDK without any sort of concurrency annotations (many basic things that
you try and do with it trigger a compiler error).

I haven't added annotations to the BatchContext things yet (because they
never leave the callback so may not need it); will address that once we
need to test those APIs.

Marking things as Sendable is consistent with the ably-cocoa public API,
which I think is going to the approach that we take threading-wise (but
will revisit in #3).