Start porting the ably-js integration tests to Swift
This is based on ably-js at 39e0677. I think that these tests will prove
to be very useful and should help reduce the testing burden when
implementing this plugin.

I wrote an example of how to map the parameterised tests into Swift
Testing, and wrote some of the Realtime-related helpers. Most of the
laborious conversion of the tests and of ObjectsHelper was then done by
Cursor, with some manual corrections. I have not looked in detail at its
output and am taking an attitude of "something is better than nothing"
towards this particular piece of work.

I will implement the remaining test cases as we implement the relevant
features.