Pass ARTRealtimeChannelInternal to plugins
This undoes a change introduced in 916bf9d, in which we needed to give
ARTRealtimeChannelInternal a reference to its public ARTRealtimeChannel,
because the LiveObjects plugin only understood the public types. This
broke the existing pattern of not referring to public types within the
internals of ably-cocoa. Here we address this by instead making the
plugins deal with ARTRealtimeChannelInternal (behind an opaque
protocol), thus also making them consistent with the internals of
ably-cocoa. This also allows us to undo the workaround introduced in
4a21e87.