Introduce status code into invalid channel state error
I didn't do this in cb427d8 because the specification hadn't yet
specified the status code (it was an outstanding question on the PR at
time of implementing), but the newly-written spec [1] for other LiveMap
getter methods _does_ specify the status code as being 400. So DRY up
the creation of these errors, and supply a status code (assuming that
the spec will be updated to specify 400 for these existing ones too).

[1] https://github.com/ably/specification/pull/341