Make callbacks Sendable
Didn't do this in ce8c022 because I didn't have a good idea of our
threading approach. But for the initial approach that we'll be taking —
namely, calling the callbacks on ably-cocoa's callback queue — I think
we'll need it.

Haven't done it for batch stuff yet because I don't yet know whether
it'll be necessary (will get a better idea when we implement it).

Now that the callback can be called on any thread, we can no longer
easily use the "capture the return value of `subscribe` pattern so that
we can unsubscribe later" pattern. So, I've decided to pass the
subscripiton as a second argument to the callback. Might be there's a
better pattern we can use (e.g. pass in an object to the `subscribe`
call, like JS `fetch`'s AbortController), but this'll do for now.