Implement the subscriptionsCallbacksScenarios
As in 8be3aed. Cursor struggled a bit when porting the JS tests, not
considering some Swift concurrency subtleties (the fact that `async let`
or `Task` tasks do not have any "synchronous part" thus needing to be
careful about not missing events, same as in c75fa2a, and the lack of
any ordering guarantees when triggering `Task`s), so I introduced the
Subscription.addListener and MainActorStorage types to perform more work
synchronously to be in line with the JS tests.