Implement applyOperationsScenarios + fixtures
This continues the porting of the JS integration tests that was started
in fa255c1, and continues the same approach and attitude taken there. It
is based on the same version of ably-js as in that commit.

I've created #60 for evaluating these ported tests for correctness and
consistency, and for bringing them in line with the latest ably-js.