Move _AblyPluginSupportPrivate to an external repository
This modifies the implementation of the plugins mechanism introduced in
7f1cc48, moving _AblyPluginSupportPrivate out of ably-cocoa and into a
new repository ably-cocoa-plugin-support [1], which becomes a shared
dependency of ably-cocoa and its plugins.

The motivation for moving _AblyPluginSupportPrivate out of ably-cocoa is
so that this library product does not appear in Xcode's "Add Package
Dependencies" UI when a user tries to add ably-cocoa to their app. This
UI's initial state offers to add all of ably-cocoa's library products,
which would likely lead to many users accidentally adding this private
library as a dependency of their app.

This change does introduce some implementation burden onto both
ably-cocoa and its plugins. In order to eliminate circular dependencies
between ably-cocoa and _AblyPluginSupportPrivate (which I needed to do
because SPM doesn't allow circular dependencies), I've had to introduce
a bunch of types in _AblyPluginSupportPrivate that duplicate ably-cocoa
types, which necessitates a load of casting and converting (basically a
bunch of boilerplate code). But, our previous usage of circular
dependencies was a bit of a hack and felt somewhat held together by
string and tape, so I'm happy for it to go away.

A pleasant side-effect of this change is that it gives us an easy way to
continue to iterate on the ably-cocoa <-> plugin internal API by
decoupling the internal API's version number from that of ably-cocoa or
the plugin; both ably-cocoa and the plugin can now declare a dependency
on the same pre-release version of ably-cocoa-plugin-support, and we can
bump this pre-release version number whenever we need to make a breaking
change. (We should still aim to move to a 1.x version of
ably-cocoa-plugin-support eventually so that users are not faced with
cryptic SPM error messages about inconsistent dependency versions when
they, say, try to upgrade ably-cocoa without upgrading the LiveObjects
plugin.)

[1] https://github.com/ably/ably-cocoa-plugin-support