Initial commit
The majority of the code here is taken from ably-cocoa at 9f1a651. I
have subsequently made the following changes, to support this library
becoming a dependency of ably-cocoa:

1. Remove all references to ably-cocoa's public types and replace them
   with equivalents that belong to this library, to eliminate circular
   dependencies.

2. Remove instance methods from APLogger and APRealtimeChannel (moving
   these to methods on APPluginAPI), so that ably-cocoa can make its
   internal ARTInternalLog and ARTRealtimeChannelInternal types conform
   to these protocols without having to worry about return or parameter
   type clashes due to the divergence introduced in 1.

3. Provide a mechanism for ably-cocoa to register an instance of
   PluginAPI (that is, the interface through which plugins access
   ably-cocoa's internals) so that it can be fetched by plugins.

The motivation for moving _AblyPluginSupportPrivate out of ably-cocoa is
so that this library product does not appear in Xcode's "Add Package
Dependencies" UI when a user tries to add ably-cocoa to their app. This
UI's initial state offers to add all of ably-cocoa's library products,
which would likely lead to many users accidentally adding this private
library as a dependency of their app.

The creation of this repository is quite a last-minute change to the
plugin architecture for ably-cocoa, with time pressure to release the
LiveObjects plugin imminently, and thus there is some technical debt
here that will need to be addressed later (documentation, CI).