Remove declaration of batch API
We have not yet implemented this and will not be doing so before our
initial release. Also the API is not fully correct — as things stand, if
you extract a LiveObject from a map inside the batch callback (see first
example on [1]) you have to use an async API to interact with it, which
is not what we want.

[1] https://ably.com/docs/liveobjects/batch