Run integration tests in serial
I've been observing a bunch of failures and hangs etc in these tests;
see the comment and the referenced issue. In order to unblock CI whilst
we prepare for a release, run the tests in serial, which seems to
improve reliability at the unfortunate cost of making them a _lot_
slower to run (from ~15s on my machine up to 4 minutes).