Change LiveMap pool delegate to return snapshot of pool
This will subsequently allow us to make more use of value types
internally.

(The change in this commit, in isolation, actually makes things a bit
messier because now we have to mock a whole ObjectsPool, with its root
object that the tests don't even need. I suppose we could instead make
the delegate just return a dictionary containing the pool's entries, but
let's see if it actually becomes a problem.)