Only allow installing via SPM
Remove support for Carthage, CocoaPods, and manual installation.

Currently just doing this so that I have something to play around with
for Swift migration but TODO tidy up and put into a v2 integration
branch. This was largely done by Claude Code.

TODO fix the version bumping scripts

TODO — something on this branch has broken Jazzy (in CI it can't even
exec it, and locally it seems to be trying to build the soak test)

Resolves #1137, resolves #2103.