Asked it to get rid of the inout error stuff that it generated
We want to be using `throws` everywhere. In hindsight, I should probably
have given the PRD a requirement like "generate the signatures that the
Swift compiler generates when it imports the Objective-C APIs"; that
would have given us maximum test compatibility. But it's a bit late for
that now; we'll probably have a bunch of stuff to fix upon trying to
build the tests.

Have not reviewed this code in detail other than noticing that the
creation of the dataEncoder had a runtime error it didn't before, but I
think it makes sense