Pin ably-cocoa-plugin-support to an exact version
Users have started getting failing builds in LiveObjects 0.1.0 because
it started resolving the 'from: "0.1.0"' ably-cocoa-plugin-support
dependency to the recently-released 0.2.0 version.

This was due to my misunderstanding of SPM dependency resolution, which
I thought used exact-version matching for 0.x versions of dependencies,
but that does not appear to be the case.