Revert priority inversion fix
This is a revert of 6b5f26b's merge of PR #1538, which fixed Xcode's
Thread Performance Checker issue of "Thread running at
QOS_CLASS_USER_INTERACTIVE waiting on a lower QoS thread running at
QOS_CLASS_DEFAULT. Investigate ways to avoid priority inversions" when
using ably-cocoa in a SwiftUI example app.

The priority inversion warning which that PR fixed was a result of the
fact that fetching +[NSRunLoop ARTSR_networkRunLoop] caused the calling
thread to block. But more recently we've started getting _another_ such
warning, this time about ARTSRProxyConnect's call to -[self.outputStream
open]. Instead of trying to make further fixes to SocketRocket (whose
code is complicated and which I'm scared of breaking), I'm going to
revert our changes and see if we can fix it in the SDK.

(That PR was originally intended as a temporary fix because we believed
that we'd be dropping SocketRocket in the near future, but that didn't
happen because our proposed replacement — using a built-in WebSocket API
— turned out not to be viable unless we increased our minimum supported
OS version, which we have not yet done. There were a few issues that we
created as a result of that PR, which I'm now undoing here. So this
resolves #1545, resolves #1546, resolves #1547.)