Add validating wrappers for dispatch_sync/async
We've had a customer report a crash that appears to be due to us trying
to submit a block to a nil queue, which results in an EXC_BAD_ACCESS
that's not very easy to interpret. Throw an exception with a helpful
message in this case instead, to help with debugging in the future.